# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/download/{event_id}": {
      "get": {
        "tags": [
          "Tickets & Invoice"
        ],
        "summary": "Download Invoice",
        "description": "Organizer Only",
        "operationId": "download_invoice_download__event_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/bookings/all-bookings": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get All Bookings",
        "description": "Fetch all bookings with:\n- Date filter (booked_date or event_date)\n- Status filter (Booked/Cancelled)\n- Sorting (total_amount/total_tickets)\n- Search by event name\n\n- Admin Only",
        "operationId": "get_all_bookings_bookings_all_bookings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "End date (YYYY-MM-DD)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by event name",
              "title": "Search"
            },
            "description": "Search by event name"
          },
          {
            "name": "date_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Dropdown option: booked_date or event_date",
              "enum": [
                "booked_date",
                "event_date"
              ],
              "default": "booked_date",
              "title": "Date Type"
            },
            "description": "Dropdown option: booked_date or event_date"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BookingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by booking status",
              "enum": [
                "booked",
                "cancelled"
              ],
              "title": "Status"
            },
            "description": "Filter by booking status"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort dropdown: total_amount or total_tickets",
              "enum": [
                "total_amount",
                "total_tickets"
              ],
              "title": "Sort By"
            },
            "description": "Sort dropdown: total_amount or total_tickets"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/get_events": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Public Events",
        "description": "\"Audience Only",
        "operationId": "get_public_events_bookings_get_events_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/bookings/available/{event_id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Available Tickets",
        "description": "\"Audience Only",
        "operationId": "get_available_tickets_bookings_available__event_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/book_event/": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Book Event",
        "description": "\"Audience Only",
        "operationId": "book_event_bookings_book_event__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/bookings/upcoming_bookings": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Upcoming Bookings",
        "description": "\"Audience Only",
        "operationId": "get_upcoming_bookings_bookings_upcoming_bookings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/bookings/past_bookings": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Upcoming Bookings",
        "description": "\"Audience Only",
        "operationId": "get_upcoming_bookings_bookings_past_bookings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/bookings/booking_cancel/{Booking_id}": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Cancel Booking",
        "operationId": "cancel_booking_bookings_booking_cancel__Booking_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/refund_status/{refund_id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Check Refund",
        "operationId": "check_refund_bookings_refund_status__refund_id__get",
        "parameters": [
          {
            "name": "refund_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Refund Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/download/{booking_id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Download Ticket",
        "description": "\"Audience Only",
        "operationId": "download_ticket_bookings_download__booking_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/backup/create": {
      "post": {
        "tags": [
          "Backup"
        ],
        "summary": "Create Backup",
        "description": "Create backup and upload to Google Drive - Admin Only",
        "operationId": "create_backup_backup_create_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/backup/list": {
      "get": {
        "tags": [
          "Backup"
        ],
        "summary": "List Backups",
        "description": "List all backups from Google Drive - Admin Only",
        "operationId": "list_backups_backup_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/backup/restore": {
      "post": {
        "tags": [
          "Backup"
        ],
        "summary": "Restore Backup",
        "description": "Download from Google Drive and restore database - Admin Only",
        "operationId": "restore_backup_backup_restore_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingCreate": {
        "properties": {
          "event_id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Event Id"
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/TicketRequest"
            },
            "type": "array",
            "title": "Tickets"
          },
          "payment_mode": {
            "type": "string",
            "enum": [
              "UPI",
              "Credit Card",
              "Debit Card"
            ],
            "title": "Payment Mode"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "tickets",
          "payment_mode"
        ],
        "title": "BookingCreate"
      },
      "BookingResponse": {
        "properties": {
          "booking_id": {
            "type": "integer",
            "title": "Booking Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/TicketSummary"
            },
            "type": "array",
            "title": "Tickets"
          },
          "total_tickets": {
            "type": "integer",
            "title": "Total Tickets"
          },
          "total_price": {
            "type": "number",
            "title": "Total Price"
          },
          "status": {
            "type": "string",
            "enum": [
              "booked",
              "cancelled"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "user_id",
          "event_id",
          "tickets",
          "total_tickets",
          "total_price",
          "status"
        ],
        "title": "BookingResponse"
      },
      "BookingStatus": {
        "type": "string",
        "enum": [
          "booked",
          "cancelled",
          "completed"
        ],
        "title": "BookingStatus"
      },
      "CancelEventRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "CancelEventRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TicketRequest": {
        "properties": {
          "ticket_type": {
            "type": "string",
            "title": "Ticket Type",
            "description": "Ticket type name (Platinum, Gold, Silver)"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Quantity",
            "description": "Number of tickets to book"
          }
        },
        "type": "object",
        "required": [
          "ticket_type",
          "quantity"
        ],
        "title": "TicketRequest"
      },
      "TicketSummary": {
        "properties": {
          "ticket_type": {
            "type": "string",
            "enum": [
              "Platinum",
              "Gold",
              "Silver"
            ],
            "title": "Ticket Type"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          }
        },
        "type": "object",
        "required": [
          "ticket_type",
          "quantity",
          "subtotal"
        ],
        "title": "TicketSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "users/login"
          }
        }
      }
    }
  }
}